+++ HW02/sergey_koleshko/.rubocop.yml
@@ -1,0 +1,8 @@
+Style/For:
+  Enabled: false
+
+Style/FrozenStringLiteralComment:
+  Enabled: false
+
+Documentation:
+  Enabled: false
