+++ HW02/sergey_koleshko/main.rb
@@ -1,23 +1,24 @@
-require './MyEnumerable.rb'
+# frozen_string_literal: true
+
+require './MyEnumerable'
 
 using MyEnumerable
-a = ["a","b","c"]
+a = %w[a b c]
 
-puts "Using my each:"
+puts 'Using my each:'
 a.my_each do |i|
-    puts i*2
+  puts i * 2
 end
-puts "Using my times:"
-5.my_times {|i| puts i}
+puts 'Using my times:'
+5.my_times { |i| puts i }
 
-puts "Using my map:"
-b = a.my_map {|i| i*2} 
+puts 'Using my map:'
+b = a.my_map { |i| i * 2 }
 puts b
 
-puts "Using my select:"
-c = a.my_select {|i| i != "a"}
+puts 'Using my select:'
+c = a.my_select { |i| i != 'a' }
 puts c
 
-puts "Using upto:"
-5.my_upto(10) {|i| puts i}
-
+puts 'Using upto:'
+5.my_upto(10) { |i| puts i }
